test(kyberlib): :white_check_mark: updating tests cases and naming convention
