+++ tests/readme.md
@@ -1,4 +1,20 @@
-# Testing
+<!-- markdownlint-disable MD033 MD041 -->
+
+<img
+src="https://kura.pro/kyberlib/images/logos/kyberlib.webp"
+alt="kyberlib's logo"
+height="261"
+width="261"
+align="right"
+/>
+
+<!-- markdownlint-enable MD033 MD041 -->
+
+# kyberlib
+
+A Robust Rust Library for CRYSTALS-Kyber Post-Quantum Cryptography.
+
+## Testing
 
 Without any feature flags `cargo test` will run through the key exchange functions and some doctests for the selected security level and mode. Running the Known Answer Tests require deterministic rng buffers from the test vector files. These files are quite large, you will need to generate them yourself. Instructions for building the KAT files are [here](./KAT/readme.md). Otherwise you can run:
 
@@ -10,13 +26,15 @@ cd KAT
 Which will clone the C reference repo, generate the KAT files, then rename and put them in the correct folder for testing.
 
 To run the known answer tests you will need to enable `KYBER_SECURITY_PARAMETERat` in `RUSTFLAGS`. To check different Kyber levels or 90's mode you will need to include those flags also. eg:
+
 ```bash
 RUSTFLAGS=' --cfg KYBER_SECURITY_PARAMETERat' cargo test --features "kyber1024 90s"
 ```
 
-For applicible x86 architectures you must export the avx2 RUSTFLAGS if you don't want to test on the reference codebase.
+For applicable x86 architectures you must export the avx2 RUSTFLAGS if you don't want to test on the reference codebase.
 
 To run a matrix of all possible features use the helper script from this folder:
+
 ```shell
 ./run_all_tests.sh
 ```
@@ -26,12 +44,12 @@ its behaviour
 
 * KAT: Runs the known answer tests
 * AVX2: Runs avx2 code on x86 platforms with compiled GAS files
-* NASM: Runs avx2 code with both GAS and NASM files seperately, requires a NASM compiler installed
+* NASM: Runs avx2 code with both GAS and NASM files separately, requires a NASM compiler installed
 
 To activate, instantiate the variables, for example:
 
 ```shell
-KAT=1 AVX2=1 NASM=1 ./run_all_tests.sh 
+KAT=1 AVX2=1 NASM=1 ./run_all_tests.sh
 ```
 
 Test files:
